// src/app/features/function-permission-group-maintenance/pages/function-settings/models/function-group-setting-edit-response.model.ts
export interface FunctionGroupSettingEditResponse {
  DBSource: string;
  ModuleId: string;
  ModuleName: string;
  Status: string;
  ParentId: string;
  OrderSeq: number;
  ButtonList: string;
  I: string;
  A: string;
  U: string;
  D: string;
  L: string;
  Selected: string;
  ModDate: string;
  ModTime: string;
  ModUser: string;
}